+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix laser pointers not working correctly when mixing UI with 2D/3D objects.
+- Fix bug in UI module where finger tips and laser pointers could not interact with
+  multiple canvases when close to each other.
 - Fix the following global input events in UxrControllerInput not being called:
   GlobalButtonStateChanged, GlobalInput1DChanged, GlobalInput2DChanged.
 - Fix warnings in example scene when loading ShotgunPump01.mp3 and ShotgunPump02.mp3